+++ rpcs3/Emu/Cell/Modules/cellPad.cpp
@@ -582,7 +582,7 @@ error_code cellPadSetActDirect(u32 port_no, vm::ptr<CellPadActParam> param)
 	if (static_cast<s32>(g_ps3_process_info.sdk_ver) > 0x1FFFFF)
 	{
 		// make sure reserved bits are 0
-		for (int i = 0; i < 6; i++)
+		for (u8 i = 0; i < 6; i++)
 		{
 			if (param->reserved[i])
 				return CELL_PAD_ERROR_INVALID_PARAMETER;
