+++ rpcs3/Emu/RSX/GL/GLProgramBuffer.h
@@ -74,7 +74,7 @@ struct GLTraits
 		{
 			// Program locations are guaranteed to not change after linking
 			// Texture locations are simply bound to the TIUs so this can be done once
-			for (int i = 0; i < rsx::limits::fragment_textures_count; ++i)
+			for (u8 i = 0; i < rsx::limits::fragment_textures_count; ++i)
 			{
 				int location;
 				if (program->uniforms.has_location(rsx::constants::fragment_texture_names[i], &location))
@@ -92,7 +92,7 @@ struct GLTraits
 				}
 			}
 
-			for (int i = 0; i < rsx::limits::vertex_textures_count; ++i)
+			for (u8 i = 0; i < rsx::limits::vertex_textures_count; ++i)
 			{
 				int location;
 				if (program->uniforms.has_location(rsx::constants::vertex_texture_names[i], &location))
